auth_service/requirements.txt
grpcio
grpcio-tools
mysql-connector-python
confluent-kafka
prometheus_client